Add first implementation of Java client
It's very early stage. I've started on the README with instructions
on how to get started but those are also pretty brief. We'll expand
on this later.

The Java client uses Gradle to build and package it, which allows us
to neatly include the compiled xgo code in the package.

The build.sh script now creates a folder per OS and copies the build
folder to the resources so it's included in the Gradle build.